+++ lib/classy.dart
@@ -1,8 +1,7 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library;
 
-export 'src/classy_data.dart';
+  export 'src/constructors/constructor.dart';
 
-// TODO: Export any libraries intended for clients of this package.
+  export 'src/constructors/from_json.dart';
+
+  export 'src/constructors/to_json.dart';