+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1-Dev
+
+- Added Constructor Macro
+- Added ToJson Macro
+- Added FromJson Macro
+
 ## 0.0.1
 
 - Initial version.